+++ lib/src/animations/indicator/water_drop_nav_painter.dart
@@ -1,6 +1,6 @@
 import 'dart:ui' show lerpDouble;
 
-import 'package:flutter/material.dart';
+import 'package:material_ui/material_ui.dart';
 
 import '../../layout/tab_slot_geometry.dart';